Fixes issues when you have more than one endpoint using a particular application that has files.
This was happen because it was attempting to download files for each *launcher*, not application. As soon as you have more than one launcher for a template, the same files will be downloaded over and over again, and because the modification dates differ for each template (they are actually the latest date of all related resources), this behavior will continue on every login.