re: #1967 - Desktop client cannot connect to logonbox server (TEST)
Required handling the redirect that /discover always returns. I do question this a little. Why is this redirect needed? It means that the client must handle it (not all do, by default, our custom netty client certainly didn't), and the location it redirects you to contains the API path, so why would you even need to look at the contents. I think /discover should not result in a redirect, but in the content.