Completed the "Stay connected" feature.
The client now detects when wireguard handshakes stop, and then also tries to contact the HTTP server to get a better idea of what the problem is. If the problem appears to be transient, i.e. the HTTP server connection fails due to no internet, or server down or whatever, it goes into the new state shown in the video "temporarily offline".  If there is no handshake, but the HTTP IS reachable, we assume it to be an invalidated configuration, and enter into the usual "Disconnected" state, with an error showing re-authorization is required.

This temporary offline state also will NOT physically shutdown and clean up the wireguard interface. This means that if say you had an active SSH connection over the VPN, but internet temporarily disappears, and then comes back, your SSH connection should spring back to life.

If there is one downside, it's how long it takes to realise that the link is temporarily down, and that is up to 3 minutes. This is because the trigger for the check being made is still dependant on the wireguard handshake. I didn't want to continuously poll the HTTP server at a faster rate than this.